Move utc_timestamp to `utils.py`
